requirements.txt
dagster==1.11.3
dagster-webserver==1.11.3
pandas==2.3.1
numpy==2.3.1
walrus-python
pydantic==2.10.6
tabulate==0.9.0